Cleaned up dialog handling using a Dialog class
The fixdialogue command will behave a tiny bit differently as it will
set participantfacing value as the dialogue would.
The previous behavior would crash if the number of boxes changed so I
could not keep it.
I kept the limitation of participants of 10 for now, does not seem like
a big deal.