Add self-made FurBW.jpg for Temple campaign
The furbw2.png file from the original mod was obviously a mix between
the Fur.jpg and Fur2.jpg of Lugaru, so I tried to redo it since furbw2.png
was sadly not licensed under CC-BY-SA 3.0.

The furwhite.png image was just the same as Fur.jpg, so I just changed
the hardcoded references to it.

Part of #26.