Use a mono font for the console text
This makes the console text more readable, as its case is quite important
and the original font uses small capital letters for lower case, which are
hard to read.

The font used is Libration Mono with height 32 and some antialiasing.
Bitmap font generated with CBFG. Liberation Mono is under the SIL Open
Font License version 1.1.

Part of #23.