CI: Rework .gitlab-ci.yml to support a CentOS 7 build
Quite a bit of the .gitlab-ci.yml file needed to be reworked so
that the commonalities between Fedora and CentOS could be leveraged
while allowing us to swap out the differences dynamically.

Thus, we can now test commits against the GCC version we use for
release builds while continuing to use latest GCC for development
builds.

As a nice side effect, there are now much fewer invocations of DNF
to install all the needed things, so CI builds should be faster!