Add FindOggVorbis.cmake module license
The copyright notice in FindOggVorbis.cmake refers to the BSD license in
a file which is currently not included in Lugaru.

Since the BSD license requires reproduction of the license notice, add
the referenced file with the included license notice.

The COPYING-CMAKE-SCRIPTS file is copied from
svn://anonsvn.kde.org/home/kde/trunk/KDE/kdemultimedia@582599

The origin of the FindOggVorbis.cmake file can be traced to
svn://anonsvn.kde.org/home/kde/trunk/KDE/kdemultimedia@529535 and the
copyright notice and license was added in
svn://anonsvn.kde.org/home/kde/trunk/KDE/kdemultimedia@582599