Remove some unneeded libs & headers
Based on some hints by the automatic dependency checking in Debian
packaging[0], it appears that, at least on Debian GNU/Linux, quite a few
of the libs that lugaru links against are not needed.

Remove linking to those libs, and use of related unneeded headers.

* For ogg/vorbis, it seems like libogg is not needed unless handling
  container metadata?

* For opengl, there seems to be only a subset of the libs that are
  required. Notably the glext.h header seems to only be needed on
  windows for GL_GENERATE_MIPMAP, but not otherwise.

* For zlib it seems to be unused, but removing zlib.h exposed a
  dependency on unistd.h which is now included directly on non-windows.

[0]
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libz.so.1 (it uses
none of the library's symbols)
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libSM.so.6 (it
uses none of the library's symbols)
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libX11.so.6 (it
uses none of the library's symbols)
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libogg.so.0 (it
uses none of the library's symbols)
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libXext.so.6 (it
uses none of the library's symbols)
dpkg-shlibdeps: warning: package could avoid a useless dependency if
debian/lugaru/usr/games/lugaru was not linked against libICE.so.6 (it
uses none of the library's symbols)