CI: Fix set installation path to work correctly for forks
The install path set for CMake assumed all the work was happening
in the master repository. However, this is is not very useful for
people who are testing changes in their own forks.

So now, we set the install path based on the project path so that
archiving works correctly.